Mock cms response in tests
