+++ app/sso/tests/test_middleware.py
@@ -6,6 +6,7 @@ import requests
 from django.urls import reverse
 
 from sso import middleware
+from navigator.tests.helpers import create_response
 
 
 def api_response_ok(*args, **kwargs):
@@ -51,11 +52,10 @@ def test_sso_middleware_api_response_ok(
 
 
 @pytest.mark.django_db
-@patch(
-    'directory_sso_api_client.client.sso_api_client.user.get_session_user',
-    api_response_bad
-)
-def test_sso_middleware_bad_response(settings, client):
+@patch('directory_sso_api_client.client.sso_api_client.user.get_session_user', api_response_bad)
+@patch('directory_cms_client.client.cms_api_client.lookup_by_slug')
+def test_sso_middleware_bad_response(mock_cms_page, settings, client):
+    mock_cms_page.return_value = create_response({'child_pages': []})
     settings.MIDDLEWARE_CLASSES = [
         'django.contrib.sessions.middleware.SessionMiddleware',
         'django.contrib.auth.middleware.AuthenticationMiddleware',
